refactor: split modules and update vite config
