+++ vite.config.ts
@@ -1,15 +1,15 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
+import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react';
 
-// https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
+  build: {
+    outDir: 'dist',
+    target: 'es2018',
+    minify: 'esbuild',
+    sourcemap: false,
+  },
   optimizeDeps: {
     exclude: ['lucide-react'],
   },
-  server: {
-    allowedHosts: [
-      '.csb.app'           
-    ]
-  }
-})
+});
